Code commented and cleaned + architecture cleaned
